create featureFormats util; remove grid format, legend, mod.
