+++ lib/layer/_layer.mjs
@@ -14,8 +14,6 @@ import distributed from './themes/distributed.mjs'
 
 import graduated from './themes/graduated.mjs'
 
-import multivariate from './themes/multivariate.mjs'
-
 export default {
   decorate,
   formats,
@@ -25,7 +23,6 @@ export default {
   themes: {
     categorized,
     distributed,
-    graduated,
-    multivariate
+    graduated
   }
 }
