+++ lib/layer/format/_format.mjs
@@ -10,8 +10,6 @@ import cluster from './cluster.mjs'
 
 import vector from './vector.mjs'
 
-import grid from './grid.mjs'
-
 export default {
   mbtiles,
   maplibre,
@@ -19,6 +17,5 @@ export default {
   mvt,
   cluster,
   geojson: vector,
-  wkt: vector,
-  grid
+  wkt: vector
 }
