+++ lib/layer/format/vector.mjs
@@ -1,52 +1,3 @@
-const formatFeatures = {
-  geojson,
-  wkt
-}
-
-const formatGeojson = new ol.format.GeoJSON
-
-const formatWKT = new ol.format.WKT
-
-function geojson(layer, features) {
-
-  return features.map((f) => {
-
-    const properties = f.properties
-
-    return new ol.Feature({
-      id: f.id,
-      geometry: formatGeojson.readGeometry(f.geometry, {
-        dataProjection: 'EPSG:' + layer.srid,
-        featureProjection: 'EPSG:' + layer.mapview.srid,
-      }),
-      ...properties
-    })
-
-  })
-}
-
-function wkt(layer, features) {
-
-  return features.map((r) => {
-
-    const properties = {}
-
-    // Assign field key and value to properties object
-    layer.fields.forEach((k, i) => properties[k] = r[i + 2])
-
-    // Return feature from geometry with properties.
-    return new ol.Feature({
-      id: r.shift(),
-      geometry: formatWKT.readGeometry(r.shift(), {
-        dataProjection: 'EPSG:' + layer.srid,
-        featureProjection: 'EPSG:' + layer.mapview.srid,
-      }),
-      ...properties
-    })
-
-  })
-}
-
 export default layer => {
 
   if (layer.properties) {
@@ -62,7 +13,7 @@ export default layer => {
     if (!features) return;
   
     let source = new ol.source.Vector({
-      features: formatFeatures[layer.format](layer, [features].flat())
+      features: mapp.utils.featureFormats[layer.format](layer, [features].flat())
     })
   
     // Geojson is a cluster layer.
@@ -93,8 +44,9 @@ export default layer => {
 
     // Create a set of feature properties for styling.
     layer.fields = [...new Set([
-      Array.isArray(layer.style.theme?.fields) ?
-        layer.style.theme.fields : layer.style.theme?.field,
+      // Array.isArray(layer.style.theme?.fields) ?
+      //   layer.style.theme.fields : layer.style.theme?.field,
+      layer.style.theme?.field,
       layer.style.label?.field,
       layer.cluster?.label
     ].flat().filter(field => !!field))]
