--- lib/layer/themes/multivariate.mjs
@@ -1,5 +1,0 @@
-export default function (theme, feature) {
-
-    console.log(feature.properties)
-
-}
