+++ lib/ui/layers/_layers.mjs
@@ -27,8 +27,6 @@ import distributed from './legends/distributed.mjs'
 
 import graduated from './legends/graduated.mjs'
 
-import grid from './legends/grid.mjs'
-
 import basic from './legends/basic.mjs'
 
 export default {
@@ -48,7 +46,6 @@ export default {
     categorized,
     distributed,
     graduated,
-    grid,
     basic,
   },
   styles: {
