+++ lib/utils/_utils.mjs
@@ -38,6 +38,8 @@ import {dataURLtoBlob} from './dataURLtoBlob.mjs'
 
 import {default as hexa} from './hexa.mjs'
 
+import * as featureFormats from './featureFormats.mjs'
+
 import loadPlugins from './loadPlugins.mjs'
 
 import getCurrentPosition from './getCurrentPosition.mjs'
@@ -71,6 +73,7 @@ export default {
   copyToClipboard,
   csvDownload,
   dataURLtoBlob,
+  featureFormats,
   gazetteer,
   getCurrentPosition,
   hexa,
